NEXT-8234 - Add snippet validator command
