+++ src/Core/System/Snippet/SnippetValidatorInterface.php
@@ -1,0 +1,8 @@
+<?php declare(strict_types=1);
+
+namespace Shopware\Core\System\Snippet;
+
+interface SnippetValidatorInterface
+{
+    public function validate(): array;
+}
