+++ src/Core/System/Test/Snippet/Filter/AddedFilterTest.php
@@ -1,6 +1,6 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Core\Framework\Test\Snippet\Filter;
+namespace Shopware\Core\System\Test\Snippet\Filter;
 
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\System\Snippet\Filter\AddedFilter;
