+++ src/Core/System/Test/Snippet/Filter/NamespaceFilterTest.php
@@ -1,38 +1,36 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Core\Framework\Test\Snippet\Filter;
+namespace Shopware\Core\System\Test\Snippet\Filter;
 
 use PHPUnit\Framework\TestCase;
-use Shopware\Core\System\Snippet\Filter\EditedFilter;
+use Shopware\Core\System\Snippet\Filter\NamespaceFilter;
 
-class EditedFilterTest extends TestCase
+class NamespaceFilterTest extends TestCase
 {
     public function testGetFilterName(): void
     {
-        static::assertSame('edited', (new EditedFilter())->getName());
+        static::assertSame('namespace', (new NamespaceFilter())->getName());
     }
 
     public function testSupports(): void
     {
-        static::assertTrue((new EditedFilter())->supports('edited'));
-        static::assertFalse((new EditedFilter())->supports(''));
-        static::assertFalse((new EditedFilter())->supports('test'));
+        static::assertTrue((new NamespaceFilter())->supports('namespace'));
+        static::assertFalse((new NamespaceFilter())->supports(''));
+        static::assertFalse((new NamespaceFilter())->supports('test'));
     }
 
-    public function testFilterOnlyCustomSnippets(): void
+    public function testFilter(): void
     {
         $snippets = [
             'firstSetId' => [
                 'snippets' => [
                     '1.bar' => [
                         'value' => '1_bar',
-                        'id' => 1,
-                        'author' => 'shopware',
+                        'translationKey' => '1.bar',
                     ],
                     '1.bas' => [
                         'value' => '1_bas',
-                        'id' => null,
-                        'author' => 'shopware',
+                        'translationKey' => '1.bas',
                     ],
                 ],
             ],
@@ -40,13 +38,11 @@ class EditedFilterTest extends TestCase
                 'snippets' => [
                     '2.bar' => [
                         'value' => '2_bar',
-                        'id' => 2,
-                        'author' => 'shopware',
+                        'translationKey' => '2.bar',
                     ],
                     '2.baz' => [
                         'value' => '2_baz',
-                        'id' => null,
-                        'author' => 'shopware',
+                        'translationKey' => '2.baz',
                     ],
                 ],
             ],
@@ -57,13 +53,89 @@ class EditedFilterTest extends TestCase
                 'snippets' => [
                     '1.bar' => [
                         'value' => '1_bar',
-                        'id' => 1,
-                        'author' => 'shopware',
+                        'translationKey' => '1.bar',
                     ],
-                    '2.bar' => [
+                    '1.bas' => [
+                        'value' => '1_bas',
+                        'translationKey' => '1.bas',
+                    ],
+                ],
+            ],
+            'secondSetId' => [
+                'snippets' => [
+                    '1.bar' => [
                         'value' => '',
                         'origin' => '',
+                        'translationKey' => '1.bar',
+                        'author' => '',
+                        'id' => null,
+                        'setId' => 'secondSetId',
+                    ],
+                    '1.bas' => [
+                        'value' => '',
+                        'origin' => '',
+                        'translationKey' => '1.bas',
+                        'author' => '',
+                        'id' => null,
+                        'setId' => 'secondSetId',
+                    ],
+                ],
+            ],
+        ];
+
+        $result = (new NamespaceFilter())->filter($snippets, ['1']);
+
+        static::assertEquals($expected, $result);
+    }
+
+    public function testFilterMultipleNamespaces(): void
+    {
+        $snippets = [
+            'firstSetId' => [
+                'snippets' => [
+                    '1.bar' => [
+                        'value' => '1_bar',
+                        'translationKey' => '1.bar',
+                    ],
+                    '1.bas' => [
+                        'value' => '1_bas',
+                        'translationKey' => '1.bas',
+                    ],
+                ],
+            ],
+            'secondSetId' => [
+                'snippets' => [
+                    '2.bar' => [
+                        'value' => '2_bar',
                         'translationKey' => '2.bar',
+                    ],
+                    '2.baz' => [
+                        'value' => '2_baz',
+                        'translationKey' => '2.baz',
+                    ],
+                    '3.foo' => [
+                        'value' => '3_foo',
+                        'translationKey' => '3.foo',
+                    ],
+                ],
+            ],
+        ];
+
+        $expected = [
+            'firstSetId' => [
+                'snippets' => [
+                    '1.bar' => [
+                        'value' => '1_bar',
+                        'translationKey' => '1.bar',
+                    ],
+                    '1.bas' => [
+                        'value' => '1_bas',
+                        'translationKey' => '1.bas',
+                    ],
+                    '3.foo' => [
+                        'value' => '',
+                        'origin' => '',
+                        'translationKey' => '3.foo',
                         'author' => '',
                         'id' => null,
                         'setId' => 'firstSetId',
@@ -80,78 +152,23 @@ class EditedFilterTest extends TestCase
                         'id' => null,
                         'setId' => 'secondSetId',
                     ],
-                    '2.bar' => [
-                        'value' => '2_bar',
-                        'id' => 2,
-                        'author' => 'shopware',
-                    ],
-                ],
-            ],
-        ];
-
-        $result = (new EditedFilter())->filter($snippets, true);
-
-        static::assertEquals($expected, $result);
-    }
-
-    public function testFilterDoesntIncludeAddedSnippets(): void
-    {
-        $snippets = [
-            'firstSetId' => [
-                'snippets' => [
-                    '1.bar' => [
-                        'value' => '1_bar',
-                        'id' => 1,
-                        'author' => 'shopware',
-                    ],
                     '1.bas' => [
-                        'value' => '1_bas',
-                        'id' => null,
-                        'author' => 'shopware',
-                    ],
-                ],
-            ],
-            'secondSetId' => [
-                'snippets' => [
-                    '2.bar' => [
-                        'value' => '2_bar',
-                        'id' => 2,
-                        'author' => 'user/admin',
-                    ],
-                    '2.baz' => [
-                        'value' => '2_baz',
-                        'id' => null,
-                        'author' => 'shopware',
-                    ],
-                ],
-            ],
-        ];
-
-        $expected = [
-            'firstSetId' => [
-                'snippets' => [
-                    '1.bar' => [
-                        'value' => '1_bar',
-                        'id' => 1,
-                        'author' => 'shopware',
-                    ],
-                ],
-            ],
-            'secondSetId' => [
-                'snippets' => [
-                    '1.bar' => [
                         'value' => '',
                         'origin' => '',
-                        'translationKey' => '1.bar',
+                        'translationKey' => '1.bas',
                         'author' => '',
                         'id' => null,
                         'setId' => 'secondSetId',
                     ],
+                    '3.foo' => [
+                        'value' => '3_foo',
+                        'translationKey' => '3.foo',
+                    ],
                 ],
             ],
         ];
 
-        $result = (new EditedFilter())->filter($snippets, true);
+        $result = (new NamespaceFilter())->filter($snippets, ['1', '3']);
 
         static::assertEquals($expected, $result);
     }
