+++ src/Core/System/Test/Snippet/Filter/TermFilterTest.php
@@ -1,25 +1,25 @@
 <?php declare(strict_types=1);
 
-namespace Shopware\Core\Framework\Test\Snippet\Filter;
+namespace Shopware\Core\System\Test\Snippet\Filter;
 
 use PHPUnit\Framework\TestCase;
-use Shopware\Core\System\Snippet\Filter\NamespaceFilter;
+use Shopware\Core\System\Snippet\Filter\TermFilter;
 
-class NamespaceFilterTest extends TestCase
+class TermFilterTest extends TestCase
 {
     public function testGetFilterName(): void
     {
-        static::assertSame('namespace', (new NamespaceFilter())->getName());
+        static::assertSame('term', (new TermFilter())->getName());
     }
 
     public function testSupports(): void
     {
-        static::assertTrue((new NamespaceFilter())->supports('namespace'));
-        static::assertFalse((new NamespaceFilter())->supports(''));
-        static::assertFalse((new NamespaceFilter())->supports('test'));
+        static::assertTrue((new TermFilter())->supports('term'));
+        static::assertFalse((new TermFilter())->supports(''));
+        static::assertFalse((new TermFilter())->supports('test'));
     }
 
-    public function testFilter(): void
+    public function testFilterWithValueMatch(): void
     {
         $snippets = [
             'firstSetId' => [
@@ -42,7 +42,7 @@ class NamespaceFilterTest extends TestCase
                     ],
                     '2.baz' => [
                         'value' => '2_baz',
-                        'translationKey' => '2.baz',
+                        'translationKey' => '2.bas',
                     ],
                 ],
             ],
@@ -55,87 +55,10 @@ class NamespaceFilterTest extends TestCase
                         'value' => '1_bar',
                         'translationKey' => '1.bar',
                     ],
-                    '1.bas' => [
-                        'value' => '1_bas',
-                        'translationKey' => '1.bas',
-                    ],
-                ],
-            ],
-            'secondSetId' => [
-                'snippets' => [
-                    '1.bar' => [
-                        'value' => '',
-                        'origin' => '',
-                        'translationKey' => '1.bar',
-                        'author' => '',
-                        'id' => null,
-                        'setId' => 'secondSetId',
-                    ],
-                    '1.bas' => [
-                        'value' => '',
-                        'origin' => '',
-                        'translationKey' => '1.bas',
-                        'author' => '',
-                        'id' => null,
-                        'setId' => 'secondSetId',
-                    ],
-                ],
-            ],
-        ];
-
-        $result = (new NamespaceFilter())->filter($snippets, ['1']);
-
-        static::assertEquals($expected, $result);
-    }
-
-    public function testFilterMultipleNamespaces(): void
-    {
-        $snippets = [
-            'firstSetId' => [
-                'snippets' => [
-                    '1.bar' => [
-                        'value' => '1_bar',
-                        'translationKey' => '1.bar',
-                    ],
-                    '1.bas' => [
-                        'value' => '1_bas',
-                        'translationKey' => '1.bas',
-                    ],
-                ],
-            ],
-            'secondSetId' => [
-                'snippets' => [
                     '2.bar' => [
-                        'value' => '2_bar',
-                        'translationKey' => '2.bar',
-                    ],
-                    '2.baz' => [
-                        'value' => '2_baz',
-                        'translationKey' => '2.baz',
-                    ],
-                    '3.foo' => [
-                        'value' => '3_foo',
-                        'translationKey' => '3.foo',
-                    ],
-                ],
-            ],
-        ];
-
-        $expected = [
-            'firstSetId' => [
-                'snippets' => [
-                    '1.bar' => [
-                        'value' => '1_bar',
-                        'translationKey' => '1.bar',
-                    ],
-                    '1.bas' => [
-                        'value' => '1_bas',
-                        'translationKey' => '1.bas',
-                    ],
-                    '3.foo' => [
                         'value' => '',
                         'origin' => '',
-                        'translationKey' => '3.foo',
+                        'translationKey' => '2.bar',
                         'author' => '',
                         'id' => null,
                         'setId' => 'firstSetId',
@@ -152,23 +75,137 @@ class NamespaceFilterTest extends TestCase
                         'id' => null,
                         'setId' => 'secondSetId',
                     ],
-                    '1.bas' => [
-                        'value' => '',
-                        'origin' => '',
-                        'translationKey' => '1.bas',
-                        'author' => '',
-                        'id' => null,
-                        'setId' => 'secondSetId',
-                    ],
-                    '3.foo' => [
-                        'value' => '3_foo',
-                        'translationKey' => '3.foo',
+                    '2.bar' => [
+                        'value' => '2_bar',
+                        'translationKey' => '2.bar',
                     ],
                 ],
             ],
         ];
 
-        $result = (new NamespaceFilter())->filter($snippets, ['1', '3']);
+        $result = (new TermFilter())->filter($snippets, '_bar');
+
+        static::assertEquals($expected, $result);
+    }
+
+    public function testFilterWithKeyMatch(): void
+    {
+        $snippets = [
+            'firstSetId' => [
+                'snippets' => [
+                    '1.bar' => [
+                        'value' => '1_bar',
+                        'translationKey' => '1.bar',
+                    ],
+                    '1.bas' => [
+                        'value' => '1_bas',
+                        'translationKey' => '1.bas',
+                    ],
+                ],
+            ],
+            'secondSetId' => [
+                'snippets' => [
+                    '2.bar' => [
+                        'value' => '2_bar',
+                        'translationKey' => '2.bar',
+                    ],
+                    '2.baz' => [
+                        'value' => '2_baz',
+                        'translationKey' => '2.bas',
+                    ],
+                ],
+            ],
+        ];
+
+        $expected = [
+            'firstSetId' => [
+                'snippets' => [
+                    '1.bar' => [
+                        'value' => '1_bar',
+                        'translationKey' => '1.bar',
+                    ],
+                    '2.bar' => [
+                        'value' => '',
+                        'origin' => '',
+                        'translationKey' => '2.bar',
+                        'author' => '',
+                        'id' => null,
+                        'setId' => 'firstSetId',
+                    ],
+                ],
+            ],
+            'secondSetId' => [
+                'snippets' => [
+                    '1.bar' => [
+                        'value' => '',
+                        'origin' => '',
+                        'translationKey' => '1.bar',
+                        'author' => '',
+                        'id' => null,
+                        'setId' => 'secondSetId',
+                    ],
+                    '2.bar' => [
+                        'value' => '2_bar',
+                        'translationKey' => '2.bar',
+                    ],
+                ],
+            ],
+        ];
+
+        $result = (new TermFilter())->filter($snippets, '.bar');
+
+        static::assertEquals($expected, $result);
+    }
+
+    public function testFilterDoesntRemoveSnippetInOtherSet(): void
+    {
+        $snippets = [
+            'firstSetId' => [
+                'snippets' => [
+                    '1.bar' => [
+                        'value' => '1_bar',
+                        'translationKey' => '1.bar',
+                    ],
+                    '1.baz' => [
+                        'value' => '1_baz',
+                        'translationKey' => '1.baz',
+                    ],
+                ],
+            ],
+            'secondSetId' => [
+                'snippets' => [
+                    '2.bar' => [
+                        'value' => '2_bar',
+                        'translationKey' => '2.bar',
+                    ],
+                    '1.baz' => [
+                        'value' => '2_baz',
+                        'translationKey' => '1.baz',
+                    ],
+                ],
+            ],
+        ];
+
+        $expected = [
+            'firstSetId' => [
+                'snippets' => [
+                    '1.baz' => [
+                        'value' => '1_baz',
+                        'translationKey' => '1.baz',
+                    ],
+                ],
+            ],
+            'secondSetId' => [
+                'snippets' => [
+                    '1.baz' => [
+                        'value' => '2_baz',
+                        'translationKey' => '1.baz',
+                    ],
+                ],
+            ],
+        ];
+
+        $result = (new TermFilter())->filter($snippets, '1_baz');
 
         static::assertEquals($expected, $result);
     }
